==> src/20 Preprocessing.ipynb <==
## Preprocessing

import pandas as pd
import numpy as np

import matplotlib as mpl
import matplotlib.pyplot as plt

save_figures = False
figure_path = "../figures/preprocessing/"
df_init = pd.read_parquet("../data/init.parquet")

### Data Cleaning

df_init

year_rename = {"2015":"Y2015",
               "2016":"Y2016"}

df_init["Year"] = df_init["Year"].cat.rename_categories(year_rename)

time_rename = {"Evening (4.30-5.30)":"Evening",
               "Morning (10.30-11.30)":"Morning",
               "Saturday Morning (11-12)":"SatMorning",
               "Saturday evening (1-2)":"SatEvening"}

df_init["Time"] = df_init["Time"].cat.rename_categories(time_rename)

df_init

df_init["ObsSize"] = df_init["ObsSize"].astype("int")
df_init["FemaleN"] = df_init["FemaleN"].astype("int")
df_init["MaleN"]   = df_init["MaleN"].astype("int")

df_init["Gender"].replace({"Group: Gender not applicable":np.nan}, inplace = True)
df_init["Age"].replace({"Group: Age not applicable":np.nan}, inplace = True)

age_rename = {"18-35":"Age_g1",
              "36 to 55":"Age_g2",
              "56+":"Age_g3"}

df_init["Age"] = df_init["Age"].cat.rename_categories(age_rename)

df_init

df_init["Child"].replace({"With child":"Child"}, inplace = True)
df_init["Child"].replace({"None":np.nan}, inplace = True)

df_init

### Interaction Terms

df_init["YearCountry"] = df_init["Year"].astype("str") + df_init["Country"].astype("str")
df_init["YearCountry"] = df_init["YearCountry"].astype("category")

df_init["YearCountryTime"] = df_init["YearCountry"].astype("str") + df_init["Time"].astype("str")
df_init["YearCountryTime"] = df_init["YearCountryTime"].astype("category")

df_init["YearCountrySupermarket"] = df_init["YearCountry"].astype("str") + df_init["Supermarket"].astype("str")
df_init["YearCountrySupermarket"] = df_init["YearCountrySupermarket"].astype("category")
df_init

df_init["YearCountryGender"] = df_init["YearCountry"].astype("str") + df_init["Gender"].astype("str")
df_init["YearCountryGender"] = df_init["YearCountryGender"].astype("category")
df_init.loc[df_init["Gender"].isna(), "YearCountryGender"] = np.nan
df_init

df_init["YearCountryAge"] = df_init["YearCountry"].astype("str") + df_init["Age"].astype("str")
df_init["YearCountryAge"] = df_init["YearCountryAge"].astype("category")
df_init.loc[df_init["Age"].isna(), "YearCountryAge"] = np.nan
df_init

### Data Preprocessing 

from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder, OrdinalEncoder, MinMaxScaler
from sklearn.model_selection import train_test_split

df_init.to_parquet("../data/before_split.parquet")

#### Test, Validation, Training 

y = df_init["y"].to_numpy()
X = df_init.loc[:, df_init.columns != "y"]

random_state = 132

# Train 
X_train, X_other, y_train, y_other = \
    train_test_split(X, y,
                     train_size = 0.8, 
                     random_state = random_state,
                     stratify = y)

# Val and Test
X_val, X_test, y_val, y_test = \
    train_test_split(X_other, y_other,
                     train_size = 0.5,
                     random_state = random_state,
                     stratify = y_other)

onehot_ftrs = ['Year', 'Country', 'Supermarket', 'Time', 
               'Weather', 'Gender', 'Age', 'Child', 
               'YearCountry', 'YearCountryTime', 'YearCountrySupermarket', 
               'YearCountryGender', 'YearCountryAge']
std_ftrs = ['ObsSize', 'FemaleN', 'MaleN']

preprocessor = ColumnTransformer(
    transformers=[
        ('onehot', OneHotEncoder(sparse=False,handle_unknown='ignore'), onehot_ftrs),
        ('std', StandardScaler(), std_ftrs)])

clf = Pipeline(steps=[('preprocessor', preprocessor)]) 

X_train_prep = clf.fit_transform(X_train)
X_val_prep = clf.transform(X_val)
X_test_prep = clf.transform(X_test)

print(X_train_prep.shape)
print(X_val_prep.shape)
print(X_test_prep.shape)

feature_names = list(preprocessor.named_transformers_["onehot"].get_feature_names(onehot_ftrs)) + \
                     preprocessor.transformers_[1][-1]

for each in feature_names:
    print(each, end = ", ")

df_train = pd.DataFrame(data = X_train_prep, 
                        columns = feature_names)
df_val = pd.DataFrame(data = X_val_prep, 
                      columns = feature_names)
df_test = pd.DataFrame(data = X_test_prep, 
                       columns = feature_names)

df_train["y"] = y_train
df_val["y"] = y_val
df_test["y"] = y_test

df_train.to_parquet("../data/prep_train.parquet")
df_val.to_parquet("../data/prep_val.parquet")
df_test.to_parquet("../data/prep_test.parquet")

#### Test, Training 

y = df_init["y"].to_numpy()
X = df_init.loc[:, df_init.columns != "y"]

random_state = 132

# Just Train and Test
X_tt_train, X_tt_test, y_tt_train, y_tt_test = \
    train_test_split(X, y,
                     train_size = 0.9, 
                     random_state = random_state,
                     stratify = y)


X_tt_train_prep = clf.fit_transform(X_tt_train)
X_tt_test_prep = clf.transform(X_tt_test)

print(X_tt_train_prep.shape)
print(X_tt_test_prep.shape)

df_tt_train = pd.DataFrame(data = X_tt_train_prep, 
                           columns = feature_names)
df_tt_test = pd.DataFrame(data = X_tt_test_prep, 
                          columns = feature_names)
df_tt_train["y"] = y_tt_train
df_tt_test["y"] = y_tt_test

df_tt_train.to_parquet("../data/prep_tt_train.parquet")
df_tt_test.to_parquet("../data/prep_tt_test.parquet")

#### All

X_all_prep = clf.fit_transform(X)
df_all = pd.DataFrame(data = X_all_prep, 
                      columns = feature_names)
print(df_all.shape)
df_all["y"] = y
df_all.to_parquet("../data/prep_all.parquet")

df_tt_test.shape
