==> src/31 Logistic Reg.ipynb <==
## Model 1: Logistic Regression

import pandas as pd
import numpy as np
import time

import matplotlib as mpl
import matplotlib.pyplot as plt

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_curve, auc, RocCurveDisplay
from sklearn.metrics import balanced_accuracy_score

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split

from sklearn.model_selection import StratifiedKFold


save_figures = False
figure_path = "../figures/models/"
# df_init = pd.read_parquet("../data/init.parquet")
# df_before = pd.read_parquet("../data/before_split.parquet")
# df_train = pd.read_parquet("../data/prep_train.parquet")
# df_val = pd.read_parquet("../data/prep_val.parquet")
# df_test = pd.read_parquet("../data/prep_test.parquet")
# df_tt_train = pd.read_parquet("../data/prep_tt_train.parquet")
# df_tt_test = pd.read_parquet("../data/prep_tt_test.parquet")
df_all = pd.read_parquet("../data/prep_all.parquet")

df_all.head()

X_all = df_all.iloc[:,0:-1].to_numpy()
y_all = df_all["y"].to_numpy()

### Logistic Regression No Penalty

random_state_LR = 132
LR_no = LogisticRegression(penalty = 'none', 
                           solver = 'saga', 
                           random_state = random_state_LR,
                           fit_intercept = False, 
                           max_iter = 10000, 
                           class_weight = 'balanced')
LR_no 

skf = StratifiedKFold(n_splits = 5, 
                      shuffle = True,
                      random_state = random_state_LR)

random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

%%time
BalAccuracy_no = []

for each_rs in random_states_split:
    start_time = time.time()
    
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    BalAccuracy_no_cv = []
    
    for train_index, val_index in skf.split(X_other_temp, y_other_temp):
        cv_start_time = time.time()

        X_train_no, X_val_no = X_other_temp[train_index], X_other_temp[val_index]
        y_train_no, y_val_no = y_other_temp[train_index], y_other_temp[val_index]

        LR_no.fit(X_train_no, y_train_no)

        y_pred_no_temp = LR_no.predict(X_val_no)
        BalAccuracy_no_cv.append(balanced_accuracy_score(y_val_no, y_pred_no_temp))

        cv_end_time = time.time()
        print(f"\tCV: {cv_end_time - cv_start_time:.3f} seconds", flush = True)
    
    BalAccuracy_no.append(np.mean(BalAccuracy_no_cv))
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_no):.4f} ({np.std(BalAccuracy_no):.4f})")

### Logistic Regression L1 Penalty

random_state_LR = 132
LR_l1 = LogisticRegression(penalty = 'l1', 
                           solver = 'saga', 
                           random_state = random_state_LR,
                           fit_intercept = False, 
                           max_iter = 10000, 
                           class_weight = 'balanced')
LR_l1

c_penalty = np.logspace(-3, 2, 21)
params_l1 = {'C':c_penalty}
random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

gridCV_l1 = GridSearchCV(estimator = LR_l1, 
                         param_grid = params_l1, 
                         scoring = 'balanced_accuracy', 
                         n_jobs = -2,
                         refit = True, 
                         cv = 5)


%%time

BalAccuracy_l1 = []

for each_rs in random_states_split:
    start_time = time.time()
    
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    gridCV_l1.fit(X_other_temp, y_other_temp)
    
    BalAccuracy_l1.append(gridCV_l1.best_score_)
    
    print(gridCV_l1.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_l1):.4f} ({np.std(BalAccuracy_l1):.4f})")


### Logistic Regression L2 Penalty

random_state_LR = 132
LR_l2 = LogisticRegression(penalty = 'l2', 
                           solver = 'saga', 
                           random_state = random_state_LR,
                           fit_intercept = False, 
                           max_iter = 10000, 
                           class_weight = 'balanced')
LR_l2

c_penalty = np.logspace(-3, 2, 21)
params_l2 = {'C':c_penalty}
random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

gridCV_l2 = GridSearchCV(estimator = LR_l2, 
                         param_grid = params_l2, 
                         scoring = 'balanced_accuracy', 
                         n_jobs = -2,
                         refit = True, 
                         cv = 5)

%%time

BalAccuracy_l2 = []

for each_rs in random_states_split:
    start_time = time.time()
    
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    gridCV_l2.fit(X_other_temp, y_other_temp)
    
    BalAccuracy_l2.append(gridCV_l2.best_score_)
    
    print(gridCV_l2.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_l2):.4f} ({np.std(BalAccuracy_l2):.4f})")


### Logistic Regression Elastic Net

random_state_LR = 132
LR_el = LogisticRegression(penalty = 'elasticnet', 
                           solver = 'saga', 
                           random_state = random_state_LR,
                           fit_intercept = False, 
                           max_iter = 10000, 
                           class_weight = 'balanced')
LR_el

c_penalty = np.logspace(-3, 2, 21)
l1_ratio = np.linspace(0.1, 0.9, 9)
params_el = {'C':c_penalty, 
             'l1_ratio':l1_ratio}
random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

gridCV_el = GridSearchCV(estimator = LR_el, 
                         param_grid = params_el, 
                         scoring = 'balanced_accuracy', 
                         n_jobs = -2,
                         refit = True, 
                         cv = 5)

%%time

BalAccuracy_el = []

for each_rs in random_states_split:
    start_time = time.time()
    
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    gridCV_el.fit(X_other_temp, y_other_temp)
    
    BalAccuracy_el.append(gridCV_el.best_score_)
    
    print(gridCV_el.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

BalAccuracy_el

print(f"{np.mean(BalAccuracy_el):.4f} ({np.std(BalAccuracy_el):.4f})")


print(f"{np.mean(BalAccuracy_no):.4f} ({np.std(BalAccuracy_no):.4f})")
print(f"{np.mean(BalAccuracy_l1):.4f} ({np.std(BalAccuracy_l1):.4f})")
print(f"{np.mean(BalAccuracy_l2):.4f} ({np.std(BalAccuracy_l2):.4f})")
print(f"{np.mean(BalAccuracy_el):.4f} ({np.std(BalAccuracy_el):.4f})")

import json

logistic_scores = {"no": BalAccuracy_no,
                   "l1": BalAccuracy_l1,
                   "l2": BalAccuracy_l2,
                   "elastic": BalAccuracy_el}

with open("../results/logistic_scores.json", "w") as outfile:
    json.dump(logistic_scores, outfile)
    
# with open("logistic.json", "r") as readfile:
#     dict_data = json.load(readfile)
