==> src/32 SVM.ipynb <==
## Model 2: Support Vector Machines

import pandas as pd
import numpy as np
import time

import matplotlib as mpl
import matplotlib.pyplot as plt

# from sklearn.linear_model import LogisticRegression
# from sklearn.metrics import roc_curve, auc, RocCurveDisplay
from sklearn.svm import SVC
from sklearn.metrics import balanced_accuracy_score
from sklearn.metrics import accuracy_score

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split

save_figures = False
figure_path = "../figures/models/"
# df_init = pd.read_parquet("../data/init.parquet")
# df_before = pd.read_parquet("../data/before_split.parquet")
# df_train = pd.read_parquet("../data/prep_train.parquet")
# df_val = pd.read_parquet("../data/prep_val.parquet")
# df_test = pd.read_parquet("../data/prep_test.parquet")
# df_tt_train = pd.read_parquet("../data/prep_tt_train.parquet")
# df_tt_test = pd.read_parquet("../data/prep_tt_test.parquet")
df_all = pd.read_parquet("../data/prep_all.parquet")

df_all.head()

X_all = df_all.iloc[:,0:-1].to_numpy()
y_all = df_all["y"].to_numpy()

X_all = np.ascontiguousarray(X_all)
y_all = np.ascontiguousarray(y_all)

### Linear Kernel, Grid Search

SVC_lin = SVC(kernel = "linear", 
              class_weight = "balanced", 
              cache_size = 1000, 
              tol = 0.002, 
              max_iter = 200_000)
SVC_lin 

c_penalty = np.logspace(-3, 2, 11)
params_lin = {'C':c_penalty}

random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

gridCV_lin = GridSearchCV(estimator = SVC_lin, 
                          param_grid = params_lin, 
                          scoring = 'balanced_accuracy', 
                          n_jobs = -2,
                          refit = True, 
                          cv = 5)
gridCV_lin

%%time

BalAccuracy_lin = []

for each_rs in random_states_split:
    start_time = time.time()
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs,
                         stratify = y_all)
    
    gridCV_lin.fit(X_other_temp, y_other_temp)
    
    BalAccuracy_lin.append(gridCV_lin.best_score_)
    
    print(gridCV_lin.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_lin):.4f} ({np.std(BalAccuracy_lin):.4f})")

### Radial Basis Function Kernel, Grid Search

SVC_rbf = SVC(kernel = "rbf", 
              class_weight = "balanced", 
              cache_size = 1000, 
              tol = 0.002, 
              max_iter = 100_000)
SVC_rbf 

c_penalty = np.logspace(-3, 2, 6)
gamma_penalty = np.logspace(-2, 2, 17)
params_rbf = {'C':c_penalty, 
              "gamma": gamma_penalty}


gridCV_rbf = GridSearchCV(estimator = SVC_rbf, 
                          param_grid = params_rbf, 
                          scoring = 'balanced_accuracy', 
                          n_jobs = -2,
                          refit = True, 
                          cv = 5, verbose = 4)
gridCV_rbf

%%time

BalAccuracy_rbf = []

for each_rs in random_states_split:
    start_time = time.time()
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    gridCV_rbf.fit(X_other_temp, y_other_temp)
    
    BalAccuracy_rbf.append(gridCV_rbf.best_score_)
    
    print(gridCV_rbf.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_rbf):.4f} ({np.std(BalAccuracy_rbf):.4f})")

print(f"{np.mean(BalAccuracy_lin):.4f} ({np.std(BalAccuracy_lin):.4f})")
print(f"{np.mean(BalAccuracy_rbf):.4f} ({np.std(BalAccuracy_rbf):.4f})")

import json

svc_scores = {"lin": BalAccuracy_lin,
              "rbf": BalAccuracy_rbf}

with open("../results/svc_scores.json", "w") as outfile:
    json.dump(svc_scores, outfile)
    
# with open("logistic.json", "r") as readfile:
#     dict_data = json.load(readfile)
