==> src/35 Boost.ipynb <==
## Model 5: Boost Methods

import pandas as pd
import numpy as np
import time

import matplotlib as mpl
import matplotlib.pyplot as plt

# from sklearn.linear_model import LogisticRegression
# from sklearn.metrics import roc_curve, auc, RocCurveDisplay
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from xgboost import XGBClassifier
from sklearn.metrics import balanced_accuracy_score
from sklearn.metrics import accuracy_score

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split

save_figures = False
figure_path = "../figures/models/"
# df_init = pd.read_parquet("../data/init.parquet")
# df_before = pd.read_parquet("../data/before_split.parquet")
# df_train = pd.read_parquet("../data/prep_train.parquet")
# df_val = pd.read_parquet("../data/prep_val.parquet")
# df_test = pd.read_parquet("../data/prep_test.parquet")
# df_tt_train = pd.read_parquet("../data/prep_tt_train.parquet")
# df_tt_test = pd.read_parquet("../data/prep_tt_test.parquet")
df_all = pd.read_parquet("../data/prep_all.parquet")

df_all.head()

X_all = df_all.iloc[:,0:-1].to_numpy()
y_all = df_all["y"].to_numpy()

### Adaboost, Grid Search

random_state = 132
Dtree = DecisionTreeClassifier()
Ada_grid = AdaBoostClassifier(base_estimator = Dtree, 
                              random_state = random_state)
Ada_grid

params_Ada = {'learning_rate': [0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 1.1],
              'n_estimators': [8, 12, 16, 20, 24, 28, 32, 36, 40],
              'base_estimator__max_depth': [1, 2, 3, 4]}

random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

gridCV_Ada = GridSearchCV(estimator = Ada_grid, 
                          param_grid = params_Ada, 
                          scoring = 'balanced_accuracy', 
                          n_jobs = -2,
                          refit = True, 
                          cv = 5, verbose = 1)
gridCV_Ada

%%time

BalAccuracy_ada = []

for each_rs in random_states_split:
    start_time = time.time()
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    n_each_class = [np.sum(y_other_temp == each_c) 
                    for each_c 
                    in np.unique(y_other_temp)]
    weight_each_class = [1/each_n 
                         for each_n 
                         in n_each_class]
    balanced_weights = np.array([weight_each_class[each_y] 
                                 for each_y 
                                 in y_other_temp])
    
    
    gridCV_Ada.fit(X_other_temp, y_other_temp, 
                   groups = None, 
                   sample_weight = balanced_weights)
    
    BalAccuracy_ada.append(gridCV_Ada.best_score_)
    
    print(gridCV_Ada.best_score_, flush = True)
    print(gridCV_Ada.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_ada):.4f} ({np.std(BalAccuracy_ada):.4f})")

### Gradient Boost, Grid Search

random_state = 132
Gradient_grid = GradientBoostingClassifier(loss = 'deviance', 
                                           random_state= random_state, 
                                           n_estimators = 50, 
                                           min_samples_split = 2)
Gradient_grid

params_Gradient = {'learning_rate': [0.05, 0.1, 0.15, 0.2, 0.25, 0.3],
                   'max_features': [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0], 
                   'max_depth': [2, 3, 4, 5, 6]}

random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

gridCV_Gradient = GridSearchCV(estimator = Gradient_grid, 
                               param_grid = params_Gradient, 
                               scoring = 'balanced_accuracy', 
                               n_jobs = -2,
                               refit = True, 
                               cv = 5, verbose = 1)
gridCV_Gradient

%%time

BalAccuracy_grad = []

for each_rs in random_states_split:
    start_time = time.time()
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    n_each_class = [np.sum(y_other_temp == each_c) 
                    for each_c 
                    in np.unique(y_other_temp)]
    weight_each_class = [1/each_n 
                         for each_n 
                         in n_each_class]
    balanced_weights = np.array([weight_each_class[each_y] 
                                 for each_y 
                                 in y_other_temp])
    
    
    gridCV_Gradient.fit(X_other_temp, y_other_temp, 
                   groups = None, 
                   sample_weight = balanced_weights)
    
    BalAccuracy_grad.append(gridCV_Gradient.best_score_)
    
    print(gridCV_Gradient.best_score_, flush = True)
    print(gridCV_Gradient.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_grad):.4f} ({np.std(BalAccuracy_grad):.4f})")

### XGB, Grid Search

random_states_split = [123, 456, 789, 
                       741, 852, 963, 
                       159, 753, 951, 357]

params_XGB = {'learning_rate': [0.25, 0.3, 0.35, 0.4, 0.45, 0.5, 0.55],
              'max_depth': [2, 3, 4, 5, 6],
              'gamma': [0.5, 0.6, 0.7, 0.8, 0.9, 1.0, 1.1,  1.2],
              'n_estimators': [10, 12, 16, 20, 24, 30, 35, 40]}


%%time

BalAccuracy_xgb = []

for each_rs in random_states_split:
    start_time = time.time()
    X_other_temp, X_test_temp, y_other_temp, y_test_temp = \
        train_test_split(X_all, y_all, 
                         test_size = 0.1, 
                         random_state = each_rs, 
                         stratify = y_all)
    
    n_each_class = [np.sum(y_other_temp == each_c) 
                    for each_c 
                    in np.unique(y_other_temp)]

    xgb_pos_scale = n_each_class[0]/n_each_class[1]
    
    
    xbg_random_state = 132
    XGB_grid = XGBClassifier(random_state= xbg_random_state, 
                             use_label_encoder=False,  
                             scale_pos_weight = xgb_pos_scale, 
                             eval_metric = 'auc')
    
    gridCV_XGB = GridSearchCV(estimator = XGB_grid, 
                           param_grid = params_XGB, 
                           scoring = 'balanced_accuracy', 
                           n_jobs = -2,
                           refit = True, 
                           cv = 5, verbose = 1)
    
    gridCV_XGB.fit(X_other_temp, y_other_temp)
    
    BalAccuracy_xgb.append(gridCV_XGB.best_score_)
    
    print(gridCV_XGB.best_params_, flush = True)
    
    end_time = time.time()
    print(f"Iteration {each_rs}: {end_time - start_time:.3f} seconds", flush = True)

print(f"{np.mean(BalAccuracy_xgb):.4f} ({np.std(BalAccuracy_xgb):.4f})")

import json

boost_scores = {"Ada": BalAccuracy_ada, 
                "Gradient": BalAccuracy_grad, 
                "XGB": BalAccuracy_xgb}

with open("../results/boost_scores.json", "w") as outfile:
    json.dump(boost_scores, outfile)
    
# with open("logistic.json", "r") as readfile:
#     dict_data = json.load(readfile)
