==> src/91 Appendix A Environment.ipynb <==
## Appendix A. Environment Summary

Environment summary for future reproducibility of the code

import sys
import importlib

print(f"Python version: {sys.version}\n")

print("Versions of each module used in this project")

modules = ["numpy", "pandas", "matplotlib", "sklearn", "xgboost", "shap"]

for each_module in modules:
    # four spaces
    print(f"    {each_module}", end = ": ")
    print(importlib.import_module(each_module).__version__)

Also, the attached [yaml file](../environment.yml) can also be used to recreate the virtual environment.
